files-exist test added
Often, it is useful to check if files exist on the board, for example,
before running certain tests, to make sure that the driver is
detecting hardware, or to check the kernel is configured
correctly.

No effort is made to ensure that a file is a character, network or block
device.

In the SE case, we want to make sure that certain device nodes exist,
for example. This test contains knowledge of SE specific files and
device nodes across boards and kernel versions.

The caller can augment the list of files by passing in the -e parameter
with a list of additional files.

The list of filenames to be checked is stored in machine and kernel
specific .lst files within the test folder.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>