TODO: ethernet test added
This test is not automated at this point. It requires the user to
confirm certain actions.

This will probably require a multinode job.

I'm thinking about splitting the test into multiple test scripts, or
into 1 script with multiple operations that run depending on the
parameters passed in, eg:

- manual ip configuration
- dhcp configuration
- ethtool settings
- mtu settings

This way, a LAVA YAML could do something like this:

- enable port
- manual ip config
- ping test from another node
- dhcp config
- ping test from another node
- ethtool set speed 100M
- ping test from another node
- ethtool set duplex half
- ping test from another node
- ethtool set duplex full
- ping test from another node
- disable port

This way, we only need a multinode ping test.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>