HACK: testplan2html: add branch support
Testing this PR, including an extra change for 'master'->'revision'

It seems this script only worked if your changes were in the 'master'
branch of the repo. By changing it to 'revision', it seems to work for
other branches.

I'm not sure what this change is doing, but Milosz suggested it during a
debugging session a while ago.

https://github.com/Linaro/test-definitions/pull/183