run multinode test on J22 on both sides
The Cambs LAVA lab network will only allow J21 to talk to J21,
and other ports cannot communicate with J21.

So when we test either the client and daemon should be on J21,
or neither should be on J21.

It's simplest if we keep the client and daemon on the same ports.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>